Adds getsCandidatesPriorities to utils
